requirements.txt
# FastAPI and Web Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0

# Data Processing and Analysis
pandas==2.3.2
numpy==2.3.3

# Machine Learning and NLP
transformers==4.56.2
torch==2.8.0
spacy==3.8.7

# Environment and Configuration
python-dotenv==1.1.1

# HTTP and Security
httpx==0.27.2
